+++ src/main/java/FlaNium/WinAPI/enums/KeyCombination.java
@@ -1,0 +1,12 @@
+package FlaNium.WinAPI.enums;
+
+public enum KeyCombination {
+
+    CTRL_A,
+    CTRL_C,
+    CTRL_V,
+    CTRL_X,
+    CTRL_A_DELETE,
+    CTRL_A_BACKSPACE;
+
+}